Align file name with stock PyTorch
