+++ src/ATen/native/xpu/LinearAlgebra.cpp
@@ -6,12 +6,15 @@
 #include <ATen/native/LinearAlgebraUtils.h>
 #include <ATen/native/ReduceOpsUtils.h>
 #include <ATen/native/utils/ParamUtils.h>
-#include <ATen/native/xpu/sycl/LinearAlgebraKernels.h>
 #include <ATen/xpu/XPUNativeFunctions.h>
+
+#include <ATen/native/xpu/sycl/LinearAlgebraKernels.h>
+#include <ATen/native/xpu/sycl/ReduceNormKernel.h>
 #include <comm/RegisterUtils.h>
 
 namespace at {
 namespace detail {
+
 static void check_linalg_norm_dtype(
     optional<ScalarType> opt_dtype,
     ScalarType self_dtype,
@@ -45,6 +48,7 @@ static void check_linalg_norm_dtype(
         ")");
   }
 }
+
 } // namespace detail
 
 Tensor& linalg_vector_norm_meta(
